moving FreeStreamInitialConditions from euler to src/physics/initial_conditions
