+++ src/testing/euler_naca0012.cpp
@@ -19,6 +19,7 @@
 #include <deal.II/fe/mapping_q.h>
 #include <deal.II/fe/mapping_manifold.h>
 
+#include "physics/initial_conditions/initial_condition.h"
 #include "physics/euler.h"
 #include "physics/manufactured_solution.h"
 #include "dg/dg_factory.hpp"
