+++ src/testing/euler_naca0012_optimization.cpp
@@ -19,6 +19,7 @@
 
 #include "euler_naca0012_optimization.hpp"
 
+#include "physics/initial_conditions/initial_condition.h"
 #include "physics/euler.h"
 #include "dg/dg_factory.hpp"
 #include "ode_solver/ode_solver_factory.h"
